# in forced_realize, unchase last op if it is upcast