Store in fp16 if it is convenient to save some memory.

Before:
```
  0 ━┳ STORE MemBuffer(idx=0, dtype=dtypes.float, st=ShapeTracker(views=(View(shape=(64, 1, 64, 112, 112, 1, 1, 1), strides=(802816, 0, 12544, 112, 1, 0, 0, 0), offset=0, mask=None, contiguous=True),)))
  1  ┗━┳ CAST (dtypes.float, False)
  2    ┗━┳ CAST (dtypes.half, False)
  3      ┗━┳ SUM ((7, 6, 5), dtypes.float)
  4        ┗━┳ CAST (dtypes.float, False)
  5          ┗━┳ MUL 
  6            ┣━━ LOAD MemBuffer(idx=1, dtype=dtypes.half, st=ShapeTracker(views=(View(shape=(1, 64, 1, 64, 4, 114, 4, 114), strides=(0, 802816, 0, 12544, 0, 112, 0, 1), offset=-113, mask=((0, 1), (0, 64), (0, 1), (0, 64), (0, 4), (1, 113), (0, 4), (1, 113)), contiguous=False), View(shape=(64, 1, 64, 112, 112, 64, 3, 3), strides=(13307904, 0, 0, 456, 1, 207936, 52440, 115), offset=0, mask=None, contiguous=False))))
  7            ┗━━ LOAD MemBuffer(idx=2, dtype=dtypes.half, st=ShapeTracker(views=(View(shape=(64, 1, 64, 112, 112, 64, 3, 3), strides=(0, 0, 576, 0, 0, 9, 3, 1), offset=0, mask=None, contiguous=False),)))
*** HSA        7 r_64_8_7_7_4_16_2_64_4_4_3_3           arg   3 mem  0.31 GB tm     12.09ms/    12.72ms ( 4896.07 GFLOPS,   25.51 GB/s)
```

```
SPLIT_REDUCEOP=0 DEFAULT_FLOAT=HALF python test/external/external_benchmark_resnet.py
layer1 slice1 x(64, 64, 112, 112)         :    367.95 ms,    1.04 tflops,    5.35 GB used,     77 kernels
layer1 slice2 x(64, 256, 112, 112)        :    361.33 ms,    0.98 tflops,    5.96 GB used,     60 kernels
layer2 slice1 x(64, 256, 56, 56)          :     71.09 ms,    2.09 tflops,    1.39 GB used,     77 kernels
layer2 slice2 x(64, 512, 28, 28)          :     37.70 ms,    2.27 tflops,    0.75 GB used,     60 kernels
layer3 slice1 x(64, 512, 28, 28)          :     47.74 ms,    3.06 tflops,    0.71 GB used,     77 kernels
layer3 slice2 x(64, 1024, 14, 14)         :     34.39 ms,    2.42 tflops,    0.39 GB used,     60 kernels
layer4 slice1 x(64, 1024, 14, 14)         :     40.34 ms,    3.62 tflops,    0.41 GB used,     77 kernels
layer4 slice2 x(64, 2048, 7, 7)           :     52.37 ms,    1.60 tflops,    0.23 GB used,     60 kernels
```

After:

```
  0 ━┳ STORE MemBuffer(idx=0, dtype=dtypes.half, st=ShapeTracker(views=(View(shape=(64, 1, 64, 112, 112, 1, 1, 1), strides=(802816, 0, 12544, 112, 1, 0, 0, 0), offset=0, mask=None, contiguous=True),)))
  1  ┗━┳ CAST (dtypes.half, False)
  2    ┗━┳ SUM ((7, 6, 5), dtypes.float)
  3      ┗━┳ CAST (dtypes.float, False)
  4        ┗━┳ MUL 
  5          ┣━━ LOAD MemBuffer(idx=1, dtype=dtypes.half, st=ShapeTracker(views=(View(shape=(1, 64, 1, 64, 4, 114, 4, 114), strides=(0, 802816, 0, 12544, 0, 112, 0, 1), offset=-113, mask=((0, 1), (0, 64), (0, 1), (0, 64), (0, 4), (1, 113), (0, 4), (1, 113)), contiguous=False), View(shape=(64, 1, 64, 112, 112, 64, 3, 3), strides=(13307904, 0, 0, 456, 1, 207936, 52440, 115), offset=0, mask=None, contiguous=False))))
  6          ┗━━ LOAD MemBuffer(idx=2, dtype=dtypes.half, st=ShapeTracker(views=(View(shape=(64, 1, 64, 112, 112, 64, 3, 3), strides=(0, 0, 576, 0, 0, 9, 3, 1), offset=0, mask=None, contiguous=False),)))
```

```
SPLIT_REDUCEOP=0 DEFAULT_FLOAT=HALF python test/external/external_benchmark_resnet.py
layer1 slice1 x(64, 64, 112, 112)         :    347.23 ms,    1.11 tflops,    4.42 GB used,     77 kernels
layer1 slice2 x(64, 256, 112, 112)        :    354.81 ms,    0.72 tflops,    5.25 GB used,     60 kernels
layer2 slice1 x(64, 256, 56, 56)          :     69.14 ms,    2.15 tflops,    1.21 GB used,     77 kernels
layer2 slice2 x(64, 512, 28, 28)          :     37.29 ms,    2.32 tflops,    0.66 GB used,     60 kernels
layer3 slice1 x(64, 512, 28, 28)          :     44.84 ms,    3.16 tflops,    0.62 GB used,     77 kernels
layer3 slice2 x(64, 1024, 14, 14)         :     34.66 ms,    2.43 tflops,    0.34 GB used,     60 kernels
layer4 slice1 x(64, 1024, 14, 14)         :     48.49 ms,    2.97 tflops,    0.37 GB used,     77 kernels
layer4 slice2 x(64, 2048, 7, 7)           :     36.50 ms,    2.34 tflops,    0.21 GB used,     60 kernels
```