@chenyuxyz 2024-04-16
i think it's fine, it moved the upcast to the next kernel and break graph differently. but i would like to see a test for this behavior so it won't be accidentally removed in the future