@chaosagent 2024-04-16
added some test, but they are very suspicious wrt multioutput/multireduce... with those, need different test cases where you know multiple kernels will be created by scheduler.