@chenyuxyz 2024-04-16
cool, very nice improvement for llm inference, llama memory access and params are basically the same now

@Qazalin fyi about the scheduling testing. given that we will have multi output and multi reduce coming, we want a testing paradigm to prevent unintended regression